requirements.txt
ecdsa==0.13
paramiko==1.16.0
progressbar==2.3
pycrypto==2.6.1
python-slugify==1.1.4
Unidecode==0.4.18
wheel==0.24.0
